Project_03_Rag/project-agent/pyproject.toml
[project]
name = "project-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.7.0",
    "gradio>=6.9.0",
    "llama-index>=0.14.16",
    "llama-index-embeddings-cohere>=0.7.0",
    "llama-index-llms-cohere>=0.7.1",
    "llama-index-llms-openai>=0.6.26",
    "llama-index-vector-stores-pinecone>=0.7.1",
    "openai>=2.26.0",
    "pinecone-client>=6.0.0",
    "pydantic>=2.12.5",
    "pyreadline3>=3.5.4",
    "python-dotenv>=1.2.2",
    "truststore>=0.10.4",
]
